Moving provider to correct namespace
